// audtech-backend/src/common/enums/status.enum.ts
export enum Status {
  ATIVO = 'ATIVO',
  INATIVO = 'INATIVO',
}

export enum StatusChecklist {
  PENDENTE = 'PENDENTE',
  EM_ANDAMENTO = 'EM_ANDAMENTO',
  CONCLUIDO = 'CONCLUIDO',
  EM_ATRASO = 'EM_ATRASO',
}

export enum StatusAcao {
  PENDENTE = 'PENDENTE',
  EM_ANDAMENTO = 'EM_ANDAMENTO',
  CONCLUIDO = 'CONCLUIDO',
  EM_ATRASO = 'EM_ATRASO',
  CANCELADO = 'CANCELADO',
}

export enum StatusNaoConformidade {
  ABERTA = 'ABERTA',
  ENCERRADA = 'ENCERRADA',
}